+++ packages/ui/VirtualList/VirtualListBasic.js
@@ -124,28 +124,12 @@ class VirtualListBasic extends Component {
 		}),
 
 		/**
-		 * Disable voice control feature of component.
+		 * An object with properties to be passed to the container DOM.
 		 *
-		 * @type {Boolean}
-		 * @public
+		 * @type {Object}
+		 * @private
 		 */
-		'data-webos-voice-disabled': PropTypes.bool,
-
-		/**
-		 * Activates the component for voice control.
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		'data-webos-voice-focused': PropTypes.bool,
-
-		/**
-		 * The voice control group label.
-		 *
-		 * @type {String}
-		 * @public
-		 */
-		'data-webos-voice-group-label': PropTypes.string,
+		containerProps: PropTypes.object,
 
 		/**
 		 * The number of items of data the list contains.
@@ -1225,7 +1209,7 @@ class VirtualListBasic extends Component {
 
 	render () {
 		const
-			{className, 'data-webos-voice-focused': voiceFocused, 'data-webos-voice-group-label': voiceGroupLabel, 'data-webos-voice-disabled': voiceDisabled, placeholderRenderer, role, style, scrollMode, ...rest} = this.props,
+			{className, containerProps, placeholderRenderer, role, style, scrollMode, ...rest} = this.props,
 			{cc, isPrimaryDirectionVertical, primary} = this,
 			scrollModeNative = scrollMode === 'native',
 			containerClasses = classNames(
@@ -1266,7 +1250,7 @@ class VirtualListBasic extends Component {
 		}
 
 		return (
-			<div className={containerClasses} data-webos-voice-focused={voiceFocused} data-webos-voice-group-label={voiceGroupLabel} data-webos-voice-disabled={voiceDisabled} ref={this.props.scrollContentRef} style={style}>
+			<div className={containerClasses} {...containerProps} ref={this.props.scrollContentRef} style={style}>
 				<div {...rest} className={contentClasses} ref={this.contentRef} role={role}>
 					{[...cc, placeholderRenderer && placeholderRenderer(primary)]}
 				</div>